Quiet warnings. Includes one actual fix: S3PutObjectRequest encodeWithCoder used the wrong method to encode an NSData.

+++ src/Amazon.S3/Model/S3MultipartUploadOperation_Internal.m
@@ -290,8 +290,8 @@ typedef struct _AWSRange {
         if(self.putRequest.data != nil)
         {
             NSRange range;
-            range.location = dataRange.location;
-            range.length = dataRange.length;
+            range.location = (NSUInteger)dataRange.location;
+            range.length = (NSUInteger)dataRange.length;
             
             self.dataForPart = [self.putRequest.data subdataWithRange:range];
         }
